Decoder layer (MLP) in decode mode works!! (#26313)
### Ticket
- #23022 

### Problem description
We do not have a Decoder layer block for DeepSeekV3.

### What's changed
Lots... lots has changed

Current result:
```
PCC for DecoderBlock in decode mode: 0.9998897722913211
```

- `reference_forwards.py`: just stores reference forward passes
(currently MLA and Decoder)
    - Can leverage function sharing here to reduce code duplication
- New decoder block
    - Different for MLP/MOE (currently only MLP decode tested)
- Before calling each module, there needs to be a `to_memory_config`
(for now...)
- Test file, very similar to MLA in terms of setting up paged config,
rope, position idxs, and doing reference forward pass
        - **Note:** only uses random weights right now
        - Does not do prefill, cache checking, or multi-iteration
- Rope is not created in using any kind of create_shared_state. This is
a TODO for another PR!
- Uses new API `add_inv_scale_to_state_dict` when testing with random
state_dict
- Reference
    - Decoder now uses `forward_mla` for attention
    - Added random weights initialization for decoder module
- MLA
    - Cleaned up RMS norm state dict creation
    - Added input reshard memory configs that can be accessed externally
- (Distributed)RMS Norm
    - Now uses float32 instead of bfloat16 when loading weights
    - input memory config for decode is inside memory config now
- config_helpers
- Inside of `get_state_dicts`, there's a torch.zeros that didn't support
the quantized datatype. I shuffled the dtype casting to make this work
- test_utils
    - Loads IO/weights as float32 if it starts off as bfloat16
- Why? Because running the reference model for MLA in bfloat16 resulted
in each matmul on cpu taking ~2min. Using float32 is much much faster,
and at the Decoder level, we want to have similar dtypes for the
underlying modules


### 🚨 TODOs After This PR 🚨 
Since we are moving at a high pace, it makes sense to merge functional
things in so that we can all work in parallel. However, this means that
there are unfinished things in this PR that need to be addressed.

- Add real weight testing
- Add multi-row testing (see how MLA does it)
- Move RoPE to a create shared
- Move the paged config/position idxs to a helper function? (where would
this go?)
- Check prefill
- Test multi-iteration decoding
- Check KV cache (maybe?)
- Move on to MOE layer testing
- Improve the PCC (a bit low to do full model bringup)
- ... and other that I missed (please update!)

### Checklist
- [ ] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/16791650843)
CI passes
- [ ] [TG DeepSeek
CI](https://github.com/tenstorrent/tt-metal/actions/runs/16791667397)
passes

---------

+++ models/demos/deepseek_v3/utils/config_helpers.py
@@ -533,7 +533,7 @@ def get_state_dicts(
         d[key].dtype == expected_dtype for d in dicts if d is not None
     ), f"Key {key} must have the dtype as {expected_dtype} in all dictionaries; instead got {[d[key].dtype if d is not None else None for d in dicts]}"
 
-    tensors = [torch.zeros(expected_shape, dtype=dtype) if d is None else d[key] for d in dicts]
+    tensors = [torch.zeros(expected_shape).to(dtype) if d is None else d[key] for d in dicts]
 
     if concat:
         return torch.concat(tensors, dim=concat_dim)
